DRY stuff out a bit
